backend/pyproject.toml

[project]
name = "todo-backend"
version = "0.1.0"
description = "Todo Backend API"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "psycopg2-binary>=2.9.10",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.12",
    "pydantic-settings>=2.6.0",
    "passlib[bcrypt]>=1.7.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
